refactor(stores): absorb FollowServiceClient into FollowStore; drop welcome locale mirror (Phase 5a)
Completes the follow-side absorption and removes a deferred mirror.

- FollowStore absorbs FollowServiceClient entirely: it now owns the @observable
  followedArtists projection, the guest follow queue, the guest-vs-authed
  branching, and the RPC calls (resolves IFollowRpcClient/IAuthService/
  IConcertStore directly), alongside its existing migration + GuestMigrationRequested
  /SignedOut handlers + receipt logic. The 4 routes that injected
  IFollowServiceClient (dashboard, discovery, my-artists, import-ticket-email)
  now inject IFollowStore. FollowServiceClient is deleted; ConcertStore keeps its
  direct storage-adapter read (no DI cycle).
- welcome-route drops the standalone @observable currentLocale mirror: the radio
  reads a getter projecting UserStore.currentLanguage (one-way checked.bind) and
  writes through changeLocale via change.trigger — the deferred task 1.4 item.

Behavior preserved: guest follow persistence (no RPC), authed RPC + cache
invalidation/rollback, the auth-boundary events, and the welcome language switch.

Refs: #368
