+++ src/services/follow-store.spec.ts
@@ -21,43 +21,30 @@ const mockLogger = {
 	}),
 }
 
+const mockAuth = { isAuthenticated: true }
+
 const mockRpcClient = {
+	listFollowed: vi.fn(async (): Promise<FollowedArtist[]> => []),
 	follow: vi.fn(async (_id: string) => undefined),
+	unfollow: vi.fn(async (_id: string) => undefined),
 	setHype: vi.fn(async (_id: string, _hype: Hype) => undefined),
 }
 
-// The delegate (FollowServiceClient) now owns the persisted guest follow queue.
-// Back it with a mutable array so per-item drain is observable, and expose the
-// guest accessors FollowStore routes through (guestFollows / removeGuestFollows
-// / clearGuestFollows).
-const mockDelegate = {
-	guestFollowsState: [] as FollowedArtist[],
-	get guestFollows(): readonly FollowedArtist[] {
-		return mockDelegate.guestFollowsState
-	},
-	followedArtists: [] as Artist[],
-	followedIds: new Set<string>(),
-	followedCount: 0,
-	hydrate: vi.fn(),
-	clear: vi.fn(() => {
-		mockDelegate.followedArtists = []
-	}),
-	// Counts calls so a test can assert the drain persisted ONCE for the batch.
-	removeGuestFollows: vi.fn((ids: readonly string[]) => {
-		const remove = new Set(ids)
-		mockDelegate.guestFollowsState = mockDelegate.guestFollowsState.filter(
-			(f) => !remove.has(f.artist.id),
-		)
-	}),
-	clearGuestFollows: vi.fn(() => {
-		mockDelegate.guestFollowsState.splice(0)
-	}),
-	follow: vi.fn(async () => undefined),
-	unfollow: vi.fn(async () => undefined),
-	listFollowed: vi.fn(async () => [] as FollowedArtist[]),
-	setHype: vi.fn(async () => undefined),
-	getFollowedArtistMap: vi.fn(async () => new Map()),
+const mockConcertStore = {
+	invalidateFollowerCache: vi.fn(),
 }
+
+// FollowStore now hydrates the guest follow queue from the guest-storage adapter
+// on construction (FollowServiceClient is dissolved). Drive it via a mutable
+// seed array that loadFollows returns and saveFollows mirrors back.
+let guestSeed: FollowedArtist[] = []
+const saveFollows = vi.fn((follows: FollowedArtist[]) => {
+	guestSeed = [...follows]
+})
+vi.mock('../adapter/storage/guest-storage', () => ({
+	loadFollows: vi.fn(() => [...guestSeed]),
+	saveFollows: (follows: FollowedArtist[]) => saveFollows(follows),
+}))
 
 const lsMap = new Map<string, string>()
 const mockStorage: ILocalStorage = {
@@ -93,14 +80,16 @@ vi.mock('aurelia', async (importOriginal) => {
 		resolve: vi.fn((token: unknown) => {
 			const map: Record<string, unknown> = {
 				ILogger: mockLogger,
+				IAuthService: mockAuth,
 				IFollowRpcClient: mockRpcClient,
-				IFollowServiceClient: mockDelegate,
+				IConcertStore: mockConcertStore,
 				ILocalStorage: mockStorage,
 				IEventAggregator: mockEa,
 			}
 			const tokenAny = token as { friendlyName?: string }
 			return map[tokenAny.friendlyName ?? ''] ?? {}
 		}),
+		observable: actual.observable,
 	}
 })
 
@@ -114,6 +103,10 @@ function makeFollow(id: string, hype: Hype = DEFAULT_HYPE): FollowedArtist {
 	return { artist: makeArtist(id, `Artist ${id}`), hype }
 }
 
+function makeFollowedArtist(id: string, name: string): FollowedArtist {
+	return { artist: makeArtist(id, name), hype: DEFAULT_HYPE }
+}
+
 describe('FollowStore', () => {
 	let sut: FollowStore
 
@@ -121,8 +114,8 @@ describe('FollowStore', () => {
 		vi.clearAllMocks()
 		subscriptions.clear()
 		lsMap.clear()
-		mockDelegate.guestFollowsState = []
-		mockDelegate.followedArtists = []
+		guestSeed = []
+		mockAuth.isAuthenticated = true
 		// clearAllMocks keeps per-test mockImplementation overrides, so restore
 		// the happy-path RPC defaults to prevent one test's failure-injection
 		// (e.g. follow throwing) from leaking into the next.
@@ -141,9 +134,174 @@ describe('FollowStore', () => {
 		return Promise.resolve()
 	}
 
+	describe('follow — cache invalidation', () => {
+		it('calls invalidateFollowerCache on successful follow', async () => {
+			mockRpcClient.follow.mockResolvedValueOnce(undefined)
+			const artist = makeArtist('a1', 'Artist One')
+
+			await sut.follow(artist)
+
+			expect(mockConcertStore.invalidateFollowerCache).toHaveBeenCalledOnce()
+		})
+
+		it('does NOT call invalidateFollowerCache when follow RPC fails', async () => {
+			mockRpcClient.follow.mockRejectedValueOnce(new Error('network error'))
+			const artist = makeArtist('a1', 'Artist One')
+
+			await expect(sut.follow(artist)).rejects.toThrow('network error')
+
+			expect(mockConcertStore.invalidateFollowerCache).not.toHaveBeenCalled()
+		})
+
+		it('rolls back the optimistic projection when the follow RPC fails', async () => {
+			mockRpcClient.follow.mockRejectedValueOnce(new Error('network error'))
+			const artist = makeArtist('a1', 'Artist One')
+
+			await expect(sut.follow(artist)).rejects.toThrow('network error')
+
+			expect(sut.followedArtists).toEqual([])
+		})
+	})
+
+	describe('unfollow — cache invalidation', () => {
+		it('calls invalidateFollowerCache on successful unfollow', async () => {
+			mockRpcClient.unfollow.mockResolvedValueOnce(undefined)
+			// Seed followedArtists so the filter doesn't no-op into empty
+			sut.followedArtists = [makeArtist('a1', 'Artist One')]
+
+			await sut.unfollow('a1')
+
+			expect(mockConcertStore.invalidateFollowerCache).toHaveBeenCalledOnce()
+		})
+
+		it('does NOT call invalidateFollowerCache when unfollow RPC fails', async () => {
+			mockRpcClient.unfollow.mockRejectedValueOnce(new Error('network error'))
+			sut.followedArtists = [makeArtist('a1', 'Artist One')]
+
+			await expect(sut.unfollow('a1')).rejects.toThrow('network error')
+
+			expect(mockConcertStore.invalidateFollowerCache).not.toHaveBeenCalled()
+		})
+	})
+
+	describe('getFollowedArtistMap — always issues RPC (hype data not cached)', () => {
+		it('calls listFollowed RPC when followedArtists is empty', async () => {
+			mockRpcClient.listFollowed.mockResolvedValueOnce([
+				makeFollowedArtist('a1', 'Artist One'),
+			])
+
+			await sut.getFollowedArtistMap()
+
+			expect(mockRpcClient.listFollowed).toHaveBeenCalledTimes(1)
+		})
+
+		it('calls listFollowed RPC on each call (hype data lives in RPC response)', async () => {
+			mockRpcClient.listFollowed.mockResolvedValue([
+				makeFollowedArtist('a1', 'Artist One'),
+			])
+			// Even with followedArtists populated, listFollowed must be called
+			// to retrieve per-artist hype values which are not stored in followedArtists.
+			sut.followedArtists = [makeArtist('a1', 'Artist One')]
+
+			await sut.getFollowedArtistMap()
+
+			expect(mockRpcClient.listFollowed).toHaveBeenCalledTimes(1)
+		})
+	})
+
+	describe('listFollowed', () => {
+		it('sets followedArtists to the mapped Artist array when authenticated', async () => {
+			mockRpcClient.listFollowed.mockResolvedValueOnce([
+				makeFollowedArtist('a1', 'Artist One'),
+				makeFollowedArtist('a2', 'Artist Two'),
+			])
+
+			await sut.listFollowed()
+
+			expect(sut.followedArtists).toEqual([
+				makeArtist('a1', 'Artist One'),
+				makeArtist('a2', 'Artist Two'),
+			])
+		})
+
+		it('sets followedArtists from guest storage follows when not authenticated', async () => {
+			mockAuth.isAuthenticated = false
+			guestSeed = [makeFollowedArtist('g1', 'Guest Artist')]
+			// Reconstruct so the guest queue hydrates from the new seed.
+			sut = new FollowStore()
+
+			await sut.listFollowed()
+
+			expect(sut.followedArtists).toEqual([makeArtist('g1', 'Guest Artist')])
+		})
+
+		it('sets followedArtists to [] when result is empty', async () => {
+			mockRpcClient.listFollowed.mockResolvedValueOnce([])
+			sut.followedArtists = [makeArtist('stale', 'Stale Artist')]
+
+			await sut.listFollowed()
+
+			expect(sut.followedArtists).toEqual([])
+		})
+	})
+
+	describe('guest follow queue (localStorage-backed, no RPC)', () => {
+		beforeEach(() => {
+			mockAuth.isAuthenticated = false
+		})
+
+		it('persists a guest follow with NO RPC and exposes it via guestFollows', async () => {
+			await sut.follow(makeArtist('g1', 'Guest One'))
+
+			expect(sut.guestFollows.map((f) => f.artist.id)).toEqual(['g1'])
+			expect(saveFollows).toHaveBeenCalled()
+			expect(mockRpcClient.follow).not.toHaveBeenCalled()
+		})
+
+		it('removeGuestFollows drains a batch in a single persist', async () => {
+			guestSeed = [
+				makeFollowedArtist('g1', 'One'),
+				makeFollowedArtist('g2', 'Two'),
+				makeFollowedArtist('g3', 'Three'),
+			]
+			sut = new FollowStore()
+
+			sut.removeGuestFollows(['g1', 'g3'])
+
+			expect(sut.guestFollows.map((f) => f.artist.id)).toEqual(['g2'])
+			expect(saveFollows).toHaveBeenCalledTimes(1)
+		})
+
+		it('clearGuestFollows empties the persisted queue', async () => {
+			guestSeed = [makeFollowedArtist('g1', 'One')]
+			sut = new FollowStore()
+
+			sut.clearGuestFollows()
+
+			expect(sut.guestFollows).toHaveLength(0)
+		})
+
+		it('setHype persists a guest hype change with NO RPC', async () => {
+			guestSeed = [makeFollowedArtist('g1', 'One')]
+			sut = new FollowStore()
+
+			await sut.setHype('g1', 'away')
+
+			expect(sut.guestFollows[0]?.hype).toBe('away')
+			expect(mockRpcClient.setHype).not.toHaveBeenCalled()
+		})
+	})
+
 	describe('GuestMigrationRequested → migrate', () => {
+		beforeEach(() => {
+			// Migration moves the guest queue to the backend, so the queue must be
+			// populated; auth state does not gate the migration RPC calls.
+			mockAuth.isAuthenticated = false
+		})
+
 		it('follows each guest artist, drains the batch, and writes the receipt', async () => {
-			mockDelegate.guestFollowsState = [makeFollow('a1'), makeFollow('a2')]
+			guestSeed = [makeFollow('a1'), makeFollow('a2')]
+			sut = new FollowStore()
 
 			await sut.migrateGuestFollows('user-1')
 
@@ -151,35 +309,26 @@ describe('FollowStore', () => {
 			expect(mockRpcClient.follow).toHaveBeenCalledWith('a1')
 			expect(mockRpcClient.follow).toHaveBeenCalledWith('a2')
 			// Batched drain emptied the queue.
-			expect(mockDelegate.guestFollowsState).toHaveLength(0)
+			expect(sut.guestFollows).toHaveLength(0)
 			// Receipt written for the account.
 			expect(lsMap.get(guestMergedReceiptKey('user-1'))).toBe('1')
 		})
 
 		it('drains the fully-migrated batch with a SINGLE persist (no O(n^2) write storm)', async () => {
-			mockDelegate.guestFollowsState = [
-				makeFollow('a1'),
-				makeFollow('a2'),
-				makeFollow('a3'),
-			]
+			guestSeed = [makeFollow('a1'), makeFollow('a2'), makeFollow('a3')]
+			sut = new FollowStore()
 
 			await sut.migrateGuestFollows('user-1')
 
-			// One batched removeGuestFollows call (single persist) rather than one
-			// write per item.
-			expect(mockDelegate.removeGuestFollows).toHaveBeenCalledOnce()
-			expect(mockDelegate.removeGuestFollows).toHaveBeenCalledWith([
-				'a1',
-				'a2',
-				'a3',
-			])
+			// One persist for the batched drain rather than one write per item.
+			// follow/setHype are guest-internal here, so the only persist on the
+			// migration path is the batched removeGuestFollows.
+			expect(saveFollows).toHaveBeenCalledOnce()
 		})
 
 		it('migrates non-default hype only for successfully followed artists', async () => {
-			mockDelegate.guestFollowsState = [
-				makeFollow('a1', 'away'),
-				makeFollow('a2'),
-			]
+			guestSeed = [makeFollow('a1', 'away'), makeFollow('a2')]
+			sut = new FollowStore()
 
 			await sut.migrateGuestFollows('user-1')
 
@@ -188,7 +337,8 @@ describe('FollowStore', () => {
 		})
 
 		it('leaves only the failed item in the queue and DEFERS the receipt', async () => {
-			mockDelegate.guestFollowsState = [makeFollow('a1'), makeFollow('a2')]
+			guestSeed = [makeFollow('a1'), makeFollow('a2')]
+			sut = new FollowStore()
 			mockRpcClient.follow.mockImplementation(async (id: string) => {
 				if (id === 'a2') throw new Error('network')
 			})
@@ -196,19 +346,15 @@ describe('FollowStore', () => {
 			await sut.migrateGuestFollows('user-1')
 
 			// a1 drained, a2 failed → remains.
-			expect(mockDelegate.guestFollowsState.map((f) => f.artist.id)).toEqual([
-				'a2',
-			])
+			expect(sut.guestFollows.map((f) => f.artist.id)).toEqual(['a2'])
 			// No receipt: the next reconcile must retry the failed item.
 			expect(lsMap.has(guestMergedReceiptKey('user-1'))).toBe(false)
 		})
 
 		it('does NOT drain an item whose SetHype fails, and DEFERS the receipt', async () => {
 			// a1 has a non-default hype; its Follow succeeds but SetHype throws.
-			mockDelegate.guestFollowsState = [
-				makeFollow('a1', 'away'),
-				makeFollow('a2'),
-			]
+			guestSeed = [makeFollow('a1', 'away'), makeFollow('a2')]
+			sut = new FollowStore()
 			mockRpcClient.setHype.mockImplementation(async (id: string) => {
 				if (id === 'a1') throw new Error('hype rpc failed')
 			})
@@ -217,16 +363,14 @@ describe('FollowStore', () => {
 
 			// a1 Follow succeeded but its hype did not migrate → keep it for retry
 			// so the non-default hype is not lost; a2 (default hype) drains.
-			expect(mockDelegate.guestFollowsState.map((f) => f.artist.id)).toEqual([
-				'a1',
-			])
-			expect(mockDelegate.removeGuestFollows).toHaveBeenCalledWith(['a2'])
+			expect(sut.guestFollows.map((f) => f.artist.id)).toEqual(['a1'])
 			// Receipt deferred so reconcile re-issues SetHype for a1.
 			expect(lsMap.has(guestMergedReceiptKey('user-1'))).toBe(false)
 		})
 
 		it('writes the receipt even when the guest queue is empty', async () => {
-			mockDelegate.guestFollowsState = []
+			guestSeed = []
+			sut = new FollowStore()
 
 			await sut.migrateGuestFollows('user-1')
 
@@ -245,7 +389,8 @@ describe('FollowStore', () => {
 		})
 
 		it('does not double-migrate: a second pass with an empty queue re-asserts the receipt only', async () => {
-			mockDelegate.guestFollowsState = [makeFollow('a1')]
+			guestSeed = [makeFollow('a1')]
+			sut = new FollowStore()
 			await sut.migrateGuestFollows('user-1')
 			expect(mockRpcClient.follow).toHaveBeenCalledTimes(1)
 
@@ -257,18 +402,17 @@ describe('FollowStore', () => {
 	})
 
 	describe('SignedOut → self-clear + cache eviction', () => {
-		it('clears the guest follow queue and evicts the projection cache via the delegate', () => {
-			mockDelegate.guestFollowsState = [makeFollow('a1')]
-			mockDelegate.followedArtists = [makeArtist('a1', 'Artist a1')]
+		it('clears the guest follow queue and evicts the projection', () => {
+			mockAuth.isAuthenticated = false
+			guestSeed = [makeFollow('a1')]
+			sut = new FollowStore()
+			sut.followedArtists = [makeArtist('a1', 'Artist a1')]
 
 			const handler = subscriptions.get(SignedOut)
 			handler?.(new SignedOut())
 
-			expect(mockDelegate.clearGuestFollows).toHaveBeenCalledOnce()
-			// Eviction routes through the delegate's own clear(), not a direct
-			// cross-object field write.
-			expect(mockDelegate.clear).toHaveBeenCalledOnce()
-			expect(mockDelegate.followedArtists).toEqual([])
+			expect(sut.guestFollows).toHaveLength(0)
+			expect(sut.followedArtists).toEqual([])
 		})
 
 		it('clears the guest-merge receipts so a fresh post-sign-out guest session can migrate again', () => {
@@ -284,12 +428,12 @@ describe('FollowStore', () => {
 			expect(sut.hasReceipt('user-2')).toBe(false)
 		})
 
-		it('clear() is idempotent and routes eviction through the delegate', () => {
+		it('clear() is idempotent and evicts the projection', () => {
+			sut.followedArtists = [makeArtist('a1', 'Artist a1')]
 			sut.clear()
 			sut.clear()
-			expect(mockDelegate.clearGuestFollows).toHaveBeenCalledTimes(2)
-			expect(mockDelegate.clear).toHaveBeenCalledTimes(2)
-			expect(mockDelegate.followedArtists).toEqual([])
+			expect(sut.guestFollows).toHaveLength(0)
+			expect(sut.followedArtists).toEqual([])
 		})
 	})
